A lot done.
I have created the environment with ambiance noise. There are hut assets that I got for free in the scene to act as the ancient village. I made "torches" using a cube and particle effects and light sources. Added material to the temple asset I got and material to the floating stone "tablets".